# contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Resume OS / JobOS Local API
  version: 1.0.0
  x-status: v1
  x-doc: docs/architecture/13-API接口设计.md
  description: >
    本地单机 REST 骨架（核心路径；全量 CRUD 于 P0 补齐，勿当完整 API）。
    仅绑定 127.0.0.1；鉴权为本地配对 token（Bearer），无 OAuth/RLS。
    WebSocket（扩展专用）契约见 ../ws/，不在本文件。
servers:
  - url: http://127.0.0.1:8000/api/v1
    description: 本机 Docker Compose
security:
  - pairingToken: []
...
components:
  securitySchemes:
    pairingToken:
      type: http
      scheme: bearer
      description: 本机首启生成的配对 token；仅接受 loopback 来源与扩展 Origin
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      schema: { type: string }
...
  responses:
    Error:
      description: 统一错误信封
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message, trace_id]
      properties:
        code:
          type: string
          enum:
            - PAIRING_TOKEN_INVALID
            - LOOPBACK_ONLY
            - EXTENSION_VERSION_MISMATCH
            - CAPTURE_RISK_HALTED
            - SCREEN_RULE_INVALID
            - SHORTLIST_CONFLICT
            - GREETING_UNSUPPORTED
            - PAYLOAD_REDLINE
            - TEMPLATE_DRIFT
            - TEMPLATE_WRONG_CHANNEL
            - TEMPLATE_LICENSE_REJECTED
            - DOCX_IMPORT_FAILED
            - CHAT_DECODE_DEGRADED
            - EVENT_ALREADY_CONFIRMED
            - RETROSPECTIVE_NOT_TERMINAL
            - VERSION_CONFLICT
            - RESUME_UNSUPPORTED_CLAIM
            - WORKFLOW_NOT_APPROVABLE
            - BUDGET_EXCEEDED
            - PAYLOAD_INVALID
            - CAPTURE_NOT_FOUND
            - EXTENSION_OFFLINE
        message: { type: string }
        trace_id: { type: string }
        details: { type: object }
    BatchRun:
      type: object
      required: [id, source, kind, status, stats, risk_halted, trigger]
      properties:
        id: { type: string, format: uuid }
        source: { type: string, enum: [boss] }
        kind: { type: string, enum: [list, detail, chat] }
        status: { type: string, enum: [queued, running, completed, aborted, risk_halted, failed] }
        stats: { type: object }
        risk_halted: { type: boolean }
        trigger: { type: string, enum: [console, extension, manual] }
        started_at: { type: [string, 'null'], format: date-time }
        finished_at: { type: [string, 'null'], format: date-time }
    CaptureCreate:
      type: object
      additionalProperties: false
      required: [kind, delay_ms]
      properties:
        kind: { type: string, enum: [list, detail] }
        query_url: { type: [string, 'null'], format: uri }
        list_params: { type: [object, 'null'] }
        max_items: { type: integer, minimum: 1, maximum: 500, default: 30 }
        ext_ids:
          type: [array, 'null']
          minItems: 1
          items: { type: string, minLength: 1 }
        detail_limit: { type: integer, minimum: 1, maximum: 100, default: 30 }
        delay_ms: { type: integer, minimum: 1800 }
    ExtensionStatus:
      oneOf:
        - type: object
          additionalProperties: false
          required: [paired]
          properties: { paired: { const: false } }
        - type: object
          additionalProperties: false
          required: [paired, instance_id, extension_version, protocol_version, connected_at]
          properties:
            paired: { const: true }
            instance_id: { type: string, format: uuid }
            extension_version: { type: string }
            protocol_version: { type: integer }
            connected_at: { type: string, format: date-time }
...
paths:
  /extension/status:
    get:
      tags: [extension]
      summary: 已鉴权控制台读取当前扩展连接状态；不返回配对令牌
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExtensionStatus' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /captures:
    post:
      tags: [capture]
      summary: 创建采集批次（人工触发；服务端下发 WS 命令给扩展）
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CaptureCreate' }
      responses:
        '202':
          description: 已排队
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchRun' }
        '403': { $ref: '#/components/responses/Error' }
    get:
      tags: [capture]
      summary: 批次列表
      responses:
        '200': { description: ok }
  /captures/{id}:
    get:
      tags: [capture]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchRun' }
        '404': { $ref: '#/components/responses/Error' }
  /captures/{id}/abort:
    post:
      tags: [capture]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: aborted }
        '409': { $ref: '#/components/responses/Error' }
...